why are we handrolling so many parsers
