--- examples/relative.txt
@@ -1,1 +1,0 @@
-a relative asset?
