+++ packages/cli-support/src/file.rs
@@ -2,8 +2,7 @@ use anyhow::Context;
 use image::{DynamicImage, EncodableLayout};
 use lightningcss::stylesheet::{MinifyOptions, ParserOptions, PrinterOptions, StyleSheet};
 use manganis_common::{
-    AssetSource, CssOptions, FileAsset, FileOptions, ImageOptions, ImageType, JsOptions,
-    JsonOptions,
+    CssOptions, FileOptions, ImageOptions, ImageType, JsOptions, JsonOptions, ResourceAsset,
 };
 use std::{
     io::{BufWriter, Write},
@@ -15,19 +14,20 @@ use swc_common::{sync::Lrc, FileName};
 use swc_common::{SourceMap, GLOBALS};
 
 pub trait Process {
-    fn process(&self, source: &AssetSource, output_path: &Path) -> anyhow::Result<()>;
+    fn process(&self, source: &ResourceAsset, output_path: &Path) -> anyhow::Result<()>;
 }
 
 /// Process a specific file asset
-pub fn process_file(file: &FileAsset, output_folder: &Path) -> anyhow::Result<()> {
-    let location = file.location();
-    let source = location.source();
-    let output_path = output_folder.join(location.unique_name());
-    file.options().process(source, &output_path)
+pub fn process_file(file: &ResourceAsset, output_folder: &Path) -> anyhow::Result<()> {
+    todo!()
+    // let location = file.location();
+    // let source = location.source();
+    // let output_path = output_folder.join(location.unique_name());
+    // file.options().process(source, &output_path)
 }
 
 impl Process for FileOptions {
-    fn process(&self, source: &AssetSource, output_path: &Path) -> anyhow::Result<()> {
+    fn process(&self, source: &ResourceAsset, output_path: &Path) -> anyhow::Result<()> {
         if output_path.exists() {
             return Ok(());
         }
@@ -61,7 +61,7 @@ impl Process for FileOptions {
 }
 
 impl Process for ImageOptions {
-    fn process(&self, source: &AssetSource, output_path: &Path) -> anyhow::Result<()> {
+    fn process(&self, source: &ResourceAsset, output_path: &Path) -> anyhow::Result<()> {
         let mut image = image::io::Reader::new(std::io::Cursor::new(&*source.read_to_bytes()?))
             .with_guessed_format()?
             .decode()?;
@@ -161,7 +161,7 @@ fn compress_png(image: DynamicImage, output_location: &Path) {
 }
 
 impl Process for CssOptions {
-    fn process(&self, source: &AssetSource, output_path: &Path) -> anyhow::Result<()> {
+    fn process(&self, source: &ResourceAsset, output_path: &Path) -> anyhow::Result<()> {
         let css = source.read_to_string()?;
 
         let css = if self.minify() { minify_css(&css) } else { css };
@@ -188,7 +188,7 @@ pub(crate) fn minify_css(css: &str) -> String {
     res.code
 }
 
-pub(crate) fn minify_js(source: &AssetSource) -> anyhow::Result<String> {
+pub(crate) fn minify_js(source: &ResourceAsset) -> anyhow::Result<String> {
     let cm = Arc::<SourceMap>::default();
 
     let js = source.read_to_string()?;
@@ -196,12 +196,13 @@ pub(crate) fn minify_js(source: &AssetSource) -> anyhow::Result<String> {
     let output = GLOBALS
         .set(&Default::default(), || {
             try_with_handler(cm.clone(), Default::default(), |handler| {
-                let filename = Lrc::new(match source {
-                    manganis_common::AssetSource::Local(path) => {
-                        FileName::Real(path.canonicalized.clone())
-                    }
-                    manganis_common::AssetSource::Remote(url) => FileName::Url(url.clone()),
-                });
+                // let filename = Lrc::new(match source {
+                //     manganis_common::ResourceAsset::Local(path) => {
+                //         FileName::Real(path.canonicalized.clone())
+                //     }
+                //     manganis_common::ResourceAsset::Remote(url) => FileName::Url(url.clone()),
+                // });
+                let filename = todo!();
                 let fm = cm.new_source_file(filename, js.to_string());
 
                 c.minify(
@@ -228,7 +229,7 @@ pub(crate) fn minify_js(source: &AssetSource) -> anyhow::Result<String> {
 }
 
 impl Process for JsOptions {
-    fn process(&self, source: &AssetSource, output_path: &Path) -> anyhow::Result<()> {
+    fn process(&self, source: &ResourceAsset, output_path: &Path) -> anyhow::Result<()> {
         let js = if self.minify() {
             minify_js(source)?
         } else {
@@ -255,7 +256,7 @@ pub(crate) fn minify_json(source: &str) -> anyhow::Result<String> {
 }
 
 impl Process for JsonOptions {
-    fn process(&self, source: &AssetSource, output_path: &Path) -> anyhow::Result<()> {
+    fn process(&self, source: &ResourceAsset, output_path: &Path) -> anyhow::Result<()> {
         let source = source.read_to_string()?;
         let json = match minify_json(&source) {
             Ok(json) => json,
