+++ packages/cli-support/src/manifest.rs
@@ -1,7 +1,7 @@
 pub use railwind::warning::Warning as TailwindWarning;
 use std::path::PathBuf;
 
-use manganis_common::{linker, AssetManifest, AssetType};
+use manganis_common::{linker, AssetType};
 
 use crate::{file::process_file, process_folder};
 
@@ -31,6 +31,65 @@ fn get_string_manganis(file: &File) -> Option<String> {
         }
     }
     None
+}
+
+/// A manifest of all assets collected from dependencies
+#[derive(Debug, PartialEq, Default, Clone)]
+pub struct AssetManifest {
+    pub(crate) assets: Vec<AssetType>,
+}
+
+impl AssetManifest {
+    /// Creates a new asset manifest
+    pub fn new(assets: Vec<AssetType>) -> Self {
+        Self { assets }
+    }
+
+    /// Returns all assets collected from dependencies
+    pub fn assets(&self) -> &Vec<AssetType> {
+        &self.assets
+    }
+
+    #[cfg(feature = "html")]
+    /// Returns the HTML that should be injected into the head of the page
+    pub fn head(&self) -> String {
+        let mut head = String::new();
+        for asset in &self.assets {
+            if let crate::AssetType::Resource(file) = asset {
+                match file.options() {
+                    crate::FileOptions::Css(css_options) => {
+                        if css_options.preload() {
+                            if let Ok(asset_path) = file.served_location() {
+                                head.push_str(&format!(
+                                    "<link rel=\"preload\" as=\"style\" href=\"{asset_path}\">\n"
+                                ))
+                            }
+                        }
+                    }
+                    crate::FileOptions::Image(image_options) => {
+                        if image_options.preload() {
+                            if let Ok(asset_path) = file.served_location() {
+                                head.push_str(&format!(
+                                    "<link rel=\"preload\" as=\"image\" href=\"{asset_path}\">\n"
+                                ))
+                            }
+                        }
+                    }
+                    crate::FileOptions::Js(js_options) => {
+                        if js_options.preload() {
+                            if let Ok(asset_path) = file.served_location() {
+                                head.push_str(&format!(
+                                    "<link rel=\"preload\" as=\"script\" href=\"{asset_path}\">\n"
+                                ))
+                            }
+                        }
+                    }
+                    _ => {}
+                }
+            }
+        }
+        head
+    }
 }
 
 /// An extension trait CLI support for the asset manifest
@@ -88,8 +147,8 @@ impl AssetManifestExt for AssetManifest {
 
         self.assets().iter().try_for_each(|asset| {
             match asset {
-                AssetType::File(file_asset) => {
-                    tracing::info!("Optimizing and bundling {}", file_asset);
+                AssetType::Resource(file_asset) => {
+                    tracing::info!("Optimizing and bundling {:?}", file_asset);
                     tracing::trace!("Copying asset from {:?} to {:?}", file_asset, location);
                     match process_file(file_asset, &location) {
                         Ok(_) => {}
@@ -98,17 +157,17 @@ impl AssetManifestExt for AssetManifest {
                             return Err(err);
                         }
                     }
+
+                    // tracing::info!("Copying folder asset {}", folder_asset);
+                    // match process_folder(folder_asset, &location) {
+                    //     Ok(_) => {}
+                    //     Err(err) => {
+                    //         tracing::error!("Failed to copy static asset: {}", err);
+                    //         return Err(err);
+                    //     }
+                    // }
                 }
-                AssetType::Folder(folder_asset) => {
-                    tracing::info!("Copying folder asset {}", folder_asset);
-                    match process_folder(folder_asset, &location) {
-                        Ok(_) => {}
-                        Err(err) => {
-                            tracing::error!("Failed to copy static asset: {}", err);
-                            return Err(err);
-                        }
-                    }
-                }
+
                 _ => {}
             }
             Ok::<(), anyhow::Error>(())
