+++ packages/common/Cargo.toml
@@ -21,8 +21,11 @@ infer = { workspace = true }
 url = { version = "2.4.0", features = ["serde"] }
 reqwest = { version = "0.12.5", features = ["blocking"] }
 tracing = "0.1.40"
+http = "1.1.0"
+http-serde = "2.1.1"
 
 [features]
+default = ["html"]
 html = []
 
 [build-dependencies]
