+++ packages/common/src/asset/tailwind.rs
@@ -1,0 +1,32 @@
+use std::{
+    fmt::Display,
+    hash::{DefaultHasher, Hash, Hasher},
+    path::{Path, PathBuf},
+};
+
+use anyhow::Context;
+use base64::Engine;
+use serde::{Deserialize, Serialize};
+use url::Url;
+
+use crate::{config, FileOptions};
+
+/// A tailwind class asset
+#[derive(Serialize, Deserialize, Debug, PartialEq, PartialOrd, Clone)]
+pub struct TailwindAsset {
+    classes: String,
+}
+
+impl TailwindAsset {
+    /// Creates a new tailwind class asset
+    pub fn new(classes: &str) -> Self {
+        Self {
+            classes: classes.to_string(),
+        }
+    }
+
+    /// Returns the classes of the tailwind class asset
+    pub fn classes(&self) -> &str {
+        &self.classes
+    }
+}
