+++ packages/common/src/lib.rs
@@ -3,13 +3,10 @@
 
 mod asset;
 mod built;
-pub mod cache;
 mod config;
 mod file;
-pub mod linker;
-mod manifest;
 
+pub mod linker;
 pub use asset::*;
 pub use config::*;
 pub use file::*;
-pub use manifest::*;
