+++ packages/common/src/linker.rs
@@ -1,8 +1,8 @@
-// code taken from https://github.com/dtolnay/linkme/,
-// MIT license
-
 //! name conventions used by the linker on different platforms.
 //! This is used to make the "link_section" magic working
+//!
+// code taken from https://github.com/dtolnay/linkme/,
+// MIT license
 
 /// Information about the manganis link section for a given platform
 #[derive(Debug, Clone, Copy)]
