+++ packages/macro/Cargo.toml
@@ -19,6 +19,7 @@ syn = { version = "2.0", features = ["full", "extra-traits"] }
 tracing-subscriber = "0.3.18"
 serde_json = "1.0"
 base64 = { workspace = true, optional = true }
+serde = { workspace = true, features = ["derive"] }
 
 manganis-common = { workspace = true }
 manganis-cli-support = { workspace = true, optional = true }
@@ -27,4 +28,5 @@ manganis-cli-support = { workspace = true, optional = true }
 manganis-common = { path = "../common", version = "0.3.0-alpha.0" }
 
 [features]
+default = ["url-encoding"]
 url-encoding = ["manganis-cli-support", "base64"]
