+++ packages/macro/src/css.rs
@@ -1,17 +1,15 @@
-use manganis_common::{
-    AssetSource, AssetType, CssOptions, FileAsset, FileOptions, ManganisSupportError,
-};
+use manganis_common::{AssetType, CssOptions, FileOptions, ManganisSupportError, ResourceAsset};
 use quote::{quote, ToTokens};
 use syn::{parenthesized, parse::Parse, LitBool};
 
-use crate::generate_link_section;
+use crate::{generate_link_section, resource::ResourceAssetParser};
 
 struct ParseCssOptions {
     options: Vec<ParseCssOption>,
 }
 
 impl ParseCssOptions {
-    fn apply_to_options(self, file: &mut FileAsset) {
+    fn apply_to_options(self, file: &mut ResourceAsset) {
         for option in self.options {
             option.apply_to_options(file);
         }
@@ -35,7 +33,7 @@ enum ParseCssOption {
 }
 
 impl ParseCssOption {
-    fn apply_to_options(self, file: &mut FileAsset) {
+    fn apply_to_options(self, file: &mut ResourceAsset) {
         match self {
             ParseCssOption::Preload(_) | ParseCssOption::Minify(_) => {
                 file.with_options_mut(|options| {
@@ -88,8 +86,7 @@ impl Parse for ParseCssOption {
 }
 
 pub struct CssAssetParser {
-    file_name: Result<String, ManganisSupportError>,
-    asset: AssetType,
+    asset: ResourceAsset,
 }
 
 impl Parse for CssAssetParser {
@@ -107,8 +104,9 @@ impl Parse for CssAssetParser {
         };
 
         let path_as_str = path.value();
-        let path: AssetSource = match AssetSource::parse_file(&path_as_str) {
-            Ok(path) => path,
+
+        let mut asset: ResourceAsset = match ResourceAsset::parse_file(&path_as_str) {
+            Ok(asset) => asset.with_options(manganis_common::FileOptions::Css(CssOptions::new())),
             Err(e) => {
                 return Err(syn::Error::new(
                     proc_macro2::Span::call_site(),
@@ -116,46 +114,17 @@ impl Parse for CssAssetParser {
                 ))
             }
         };
-        let mut this_file = FileAsset::new(path.clone())
-            .with_options(manganis_common::FileOptions::Css(CssOptions::new()));
+
         if let Some(parsed_options) = parsed_options {
-            parsed_options.apply_to_options(&mut this_file);
+            parsed_options.apply_to_options(&mut asset);
         }
 
-        let asset = manganis_common::AssetType::File(this_file.clone());
-
-        let file_name = if this_file.url_encoded() {
-            #[cfg(not(feature = "url-encoding"))]
-            return Err(syn::Error::new(
-                proc_macro2::Span::call_site(),
-                "URL encoding is not enabled. Enable the url-encoding feature to use this feature",
-            ));
-            #[cfg(feature = "url-encoding")]
-            Ok(crate::url_encoded_asset(&this_file).map_err(|e| {
-                syn::Error::new(
-                    proc_macro2::Span::call_site(),
-                    format!("Failed to encode file: {}", e),
-                )
-            })?)
-        } else {
-            this_file.served_location()
-        };
-
-        Ok(CssAssetParser { file_name, asset })
+        Ok(CssAssetParser { asset })
     }
 }
 
 impl ToTokens for CssAssetParser {
     fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
-        let file_name = crate::quote_path(&self.file_name);
-
-        let link_section = generate_link_section(self.asset.clone());
-
-        tokens.extend(quote! {
-            {
-                #link_section
-                #file_name
-            }
-        })
+        ResourceAssetParser::to_ref_tokens(&self.asset, tokens)
     }
 }
