+++ packages/macro/src/file.rs
@@ -1,50 +1,26 @@
-use manganis_common::{AssetSource, AssetType, FileAsset, ManganisSupportError};
+use manganis_common::{AssetType, ManganisSupportError, ResourceAsset};
 use quote::{quote, ToTokens};
 use syn::{parenthesized, parse::Parse};
 
-use crate::generate_link_section;
+use crate::{generate_link_section, resource::ResourceAssetParser};
 
 pub struct FileAssetParser {
-    file_name: Result<String, ManganisSupportError>,
-    asset: AssetType,
+    pub asset: ResourceAssetParser,
 }
 
 impl Parse for FileAssetParser {
     fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
         let inside;
         parenthesized!(inside in input);
-        let path = inside.parse::<syn::LitStr>()?;
 
-        let path_as_str = path.value();
-        let path = match AssetSource::parse_file(&path_as_str) {
-            Ok(path) => path,
-            Err(e) => {
-                return Err(syn::Error::new(
-                    proc_macro2::Span::call_site(),
-                    format!("{e}"),
-                ))
-            }
-        };
-        let this_file = FileAsset::new(path);
-        let asset = manganis_common::AssetType::File(this_file.clone());
+        let asset = inside.parse::<ResourceAssetParser>()?;
 
-        let file_name = this_file.served_location();
-
-        Ok(FileAssetParser { file_name, asset })
+        Ok(FileAssetParser { asset })
     }
 }
 
 impl ToTokens for FileAssetParser {
     fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
-        let file_name = crate::quote_path(&self.file_name);
-
-        let link_section = generate_link_section(self.asset.clone());
-
-        tokens.extend(quote! {
-            {
-                #link_section
-                #file_name
-            }
-        })
+        self.asset.to_tokens(tokens)
     }
 }
