+++ packages/manganis/Cargo.toml
@@ -14,6 +14,17 @@ keywords = ["assets"]
 
 [dependencies]
 manganis-macro = { workspace = true, optional = true }
+once_cell = "1.19.0"
+dirs = "5.0.1"
+infer = { workspace = true }
+manganis-common = { workspace = true }
+dunce = "1.0.2"
+
+[target.'cfg(target_os = "macos")'.dependencies]
+core-foundation = "0.9.3"
+
+[target.'cfg(target_os ="macos")'.dependencies]
+
 
 [features]
 default = ["macro"]
