+++ packages/resolver/Cargo.toml
@@ -6,3 +6,10 @@ version.workspace = true
 [dependencies]
 infer = { workspace = true }
 manganis-common = { workspace = true }
+manganis = { workspace = true }
+dunce = "1.0.2"
+
+[target.'cfg(target_os = "macos")'.dependencies]
+core-foundation = "0.9.3"
+
+[target.'cfg(target_os ="macos")'.dependencies]
