+++ packages/virtual-file/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "virtual-file"
+edition = "2021"
+version.workspace = true
+
+[dependencies]
